pyproject.toml
[tool.poetry]
name = "pypropel"
version = "0.1.0"
description = ""
authors = ["Jianfeng Sun <jianfeng.sunmt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.0.1"
pandas = "^2.2.2"
biopython = "^1.84"
seaborn = "^0.13.2"
matplotlib = "^3.9.1"
scikit-learn = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
